test(debug-info): move DWARF e2e tests to a dedicated module with typed JSON I/O
The DWARF line-table tests pin pipeline properties, not CLI behavior, so
they get their own integration test category next to the CLI tests. All
standard JSON input construction and output parsing now goes through the
typed solx-standard-json structures instead of serde_json::Value string
surgery; the bytecode invariance test derives both compilations from one
fixture via two typed output selections, keeping "differ only in the
debugInfo selection" true by construction.
